Refactoring Curate::User
Attempting to keep the concerns "mixable" in that we may have a
curate instance in which the developers chooses don't want the
person attributes
