+++ spec/models/curate/user_spec.rb
